backend/requirements.txt
fastapi
uvicorn[standard]
deepgram-sdk
openai
supabase-py
aioredis
python-dotenv
